start to modify daomem into dao jdbc
